Original sine wave file + error for sine_wave.py
Contrary to the Issue post, sine_wave.py actually does not have the saving feature hashtagged out. If you run it while possessing a writer like ffmpeg, you may get the error message. If you hashtag out the saving with the player, then you should just have plt.show and see a sine wave across the screen. But it will not save as a video file.